requirements.txt
biopython>=1.78
matplotlib>=3.3.0
numpy>=1.19.0
ipywidgets>=7.5.0